move User Model in to User Folder
