+++ app/Http/Controllers/Api/V1/UserController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Api\V1;
 
 use App\Models\Task;
-use App\Models\User;
+use App\Models\User\User;
 use App\Repositories\UserRepository;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
