+++ app/Http/Controllers/SocialAuthController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Profile;
-use App\Models\User;
+use App\Models\User\User;
 use Illuminate\Http\Request;
 use JWTAuth;
 use Socialite;
