+++ app/Models/Configuration/Traits/Scope/ConfigurationScope.php
