+++ app/Models/Profile/Traits/Scope/ProfileScope.php
