+++ app/Models/User/Traits/Scope/UserScope.php
