+++ app/Repositories/Backend/UserRepository.php
@@ -2,7 +2,7 @@
 
 namespace App\Repositories\Backend;
 
-use App\Models\User;
+use App\Models\User\User;
 use App\Repositories\BaseRepository;
 
 /**
