+++ config/jwt.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Models\User;
+use App\Models\User\User;
 
 /*
  * This file is part of jwt-auth.
