+++ database/factories/UserFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Models\User;
+use App\Models\User\User;
 use Faker\Generator as Faker;
 
 /*
